refactor(staking): extract LivenessTracker + DoubleSignDetector
slashing.rs was 954 LOC carrying three independent concerns:

- LivenessTracker (sliding-window downtime detector) +
  LIVENESS_WINDOW / MIN_SIGNED_PER_WINDOW / DOWNTIME_SLASH_BP /
  DOWNTIME_JAIL_BLOCKS constants
- DoubleSignEvidence + DoubleSignDetector + DOUBLE_SIGN_SLASH_BP
- SlashingEngine (the orchestrator) + tests

Pulled out the first two into slashing/{liveness,double_sign}.rs.
slashing.rs drops to 728 LOC, focused on SlashingEngine + tests.

Pure mechanical extraction — public API at sentrix_staking::slashing::*
unchanged via re-exports.

Phase 3 follow-on of the workspace audit.
